backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.74.1
fastapi-login==1.7.3
h11==0.13.0
idna==3.3
mongoengine==0.24.0
passlib==1.7.4
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.5
